Spread season totals over the weeks they are really played; refresh 2027 settings
Weekly lines divided season totals by the regular-season periods, but
production runs through the fantasy playoffs: 23.4 effective weeks in every
season measured, so lines were 23% high. pool.effective_weeks measures it.
Re-ingesting 2027 picked up the centre limit of 3.

+++ app/draft/optimizer.py
@@ -79,7 +79,7 @@ def candidates_from(
     projections: Sequence[PlayerProjection],
     board: PriceBoard,
     *,
-    periods: int,
+    periods: float,
     availability: float,
     keys: Iterable[str],
 ) -> list[Candidate]:
@@ -89,7 +89,7 @@ def candidates_from(
     model already decided they will not be rostered.
     """
     wanted = tuple(keys)
-    divisor = max(1, periods)
+    divisor = max(1.0, periods)
     out: list[Candidate] = []
     for projection in projections:
         price = board.price_of(projection.player_id)
